vision_opportunity/requirements.txt
rich>=13.0.0
click>=8.1.0
pydantic>=2.0.0
jsonschema>=4.17.0
typing-extensions>=4.5.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.28.0 